Signed-off-by: CachorroC <juankpato87@gmail.com>
a Please enter the commit message for your changes. Lines starting
a with 'a' will be ignored, and an empty message aborts the commit.
a
a On branch main
a Your branch is up to date with 'origin/main'.
a
a Changes to be committed:
a	modified:   .dockerignore
a	modified:   .env
a